Fix sort ordering in list and search views
Formerly, we were sorting lexicographically by id in the list
view (thanks, javascript!) and were missing dates for a lot of the items
in the search view, leading to no apparent order.  This commit ensures
we have a date field everywhere, that we sort by it in the search, and
that we fall back to numerically sorted id when there's no date, or in
the plain list view.

Fixes #85.