Implement plasmid mapper in scala
The plasmid map was pretty separable from everything else, so I ported it over
first.  I changed the API so that instead of giving you the map for a plasmid in
the database, it's a standalone service where you just POST text containing the
plasmid data and it gives you back the map.

Upon reflection, the data format the map endpoint returned was awful (why group
the features by name?  why include the total count of point features of that
name on each point feature?  why use different names for position in point and
regional features?).  Thus, I changed it to something hopefully more sensible.
Since I didn't want to get into the hairy mess of rewriting the plasmid mapper
frontend yet, I wrote a js adapter that takes the new data format and converts
it to the old one before passing it to the plasmid mapper frontend code.

Also when writing tests, I found a bug in the old code where if you had a
plasmid whose length wasn't evenly divisible by 3, you'd get two copies of the
protein features, off by 1 or 2 bases depending on the frame shift.  Yay tests!