Set up an alternate header-based auth method
In order to transition endpoints gradually from the ruby app to the
scala one, we need some sort of shared authentication method.  This
commit is the ruby-side changes needed to make this happen.  In the new
flow, auth with google happens in scala, and then we send signed
headers containing the user id to the ruby application.  If the
signature is valid and sufficiently recent, we accept that as the
authenticated user id.