Fix an issue where the "new item" would no longer autofill user
The function that determines the current user_id wasn't being used for
autofilling the user when creating a new object; rather it was reaching
directly into the session (which we're no longer filling when the
request is proxied).

The solution is to call the curr_uid function from the Authorization
module directly.