Fix a bug when linking two items, one of which is a prefix of the other.
When auto-linking items, we correctly identify the full mention of another
item by regex. However, when substituting in the links, we then do a simple
text substitution of the full match. This is a problem if a single entry
mentions two items, one of which is a prefix of the other (e.g. ASP2 and
ASP20), as we'll then substitute out the prefix on both of them and fail to
link the later entry to the correct place.

To fix, this commit just uses a regex with negative lookahead for the
substitution to ensure that there's not a digit immediately following
anything we're substituting.