Fix the plasmid map autoconversion from coffeescript
Summary:

In a previous commit, I converted the plasmid map code from coffeescript to
typescript with an automated tool. However, this left it not completely
functional. This commit refactors the plasmid map so that it now works again,
and cleaned up some weird and/or nonidiomatic design choices there too. Still
lots more cleanup to do on the plasmid map in the future, though