Show antibiotics on the strain, combining intrinsics and plasmid resistances
Summary:

Previously, only the plasmids table listed antibiotic resistances. However,
this is both inconvenient to look up if you're looking at the strain page, and
potentially wrong if the strain has an in-chromosome resistance that can't be
specified anywhere.

This commit adds an intrinsic resistances field to the bacteria table and
combines the resistances from plasmids and the strain itself in the UI.