Fix a bug leading to exponential amplification of links
When an item links to another item multiple times, there was a bug such
that it would replace recursively, such that each time, it substitutes
all previous matches. This wouldn't be so bad, except that the
subsitution makes the name appear twice: once in the href, and once in
the text. This leads to exponential amplification so that if you have N
occurences in the original text, you end up with 2**N subsitutions.

With one particularly bad case where N=30, this crashed the database on
any load of that item, or any search that would return that as a result.